+++ graphql/test/schema_test.ml
@@ -106,6 +106,26 @@ let suite = [
       ]
     ])
   );
+  ("undefined field on query root", `Quick, fun () ->
+    let query = "{ foo { bar } }" in
+    test_query query (`Assoc [
+      "errors", `List [
+        `Assoc [
+          "message", `String "Field 'foo' is not defined on type 'query'"
+        ]
+      ]
+    ])
+  );
+  ("undefined field on object type", `Quick, fun () ->
+    let query = "{ users { id foo } }" in
+    test_query query (`Assoc [
+      "errors", `List [
+        `Assoc [
+          "message", `String "Field 'foo' is not defined on type 'user'"
+        ]
+      ]
+    ])
+  );
   ("fragments cannot form cycles", `Quick, fun () ->
     let query = "
       fragment F1 on Foo {
@@ -131,5 +151,107 @@ let suite = [
         ]
       ]
     ])
+  );
+  ("introspection query should be accepted", `Quick, fun () ->
+    let query = "
+      query IntrospectionQuery {
+        __schema {
+          queryType { name }
+          mutationType { name }
+          subscriptionType { name }
+          types {
+            ...FullType
+          }
+          directives {
+            name
+            description
+            locations
+            args {
+              ...InputValue
+            }
+          }
+        }
+      }
+
+      fragment FullType on __Type {
+        kind
+        name
+        description
+        fields(includeDeprecated: true) {
+          name
+          description
+          args {
+            ...InputValue
+          }
+          type {
+            ...TypeRef
+          }
+          isDeprecated
+          deprecationReason
+        }
+        inputFields {
+          ...InputValue
+        }
+        interfaces {
+          ...TypeRef
+        }
+        enumValues(includeDeprecated: true) {
+          name
+          description
+          isDeprecated
+          deprecationReason
+        }
+        possibleTypes {
+          ...TypeRef
+        }
+      }
+
+      fragment InputValue on __InputValue {
+        name
+        description
+        type { ...TypeRef }
+        defaultValue
+      }
+
+      fragment TypeRef on __Type {
+        kind
+        name
+        ofType {
+          kind
+          name
+          ofType {
+            kind
+            name
+            ofType {
+              kind
+              name
+              ofType {
+                kind
+                name
+                ofType {
+                  kind
+                  name
+                  ofType {
+                    kind
+                    name
+                    ofType {
+                      kind
+                      name
+                    }
+                  }
+                }
+              }
+            }
+          }
+        }
+      }
+    " in
+    match Graphql_parser.parse query with
+    | Error err -> failwith err
+    | Ok doc ->
+        begin match Graphql.Schema.execute Test_schema.schema () doc with
+        | Ok _ -> ()
+        | Error err -> failwith (Yojson.Basic.pretty_to_string err)
+        end
   );
 ]