+++ graphql/test/test.ml
@@ -4,5 +4,6 @@ let () =
     "arguments", Argument_test.suite;
     "variables", Variable_test.suite;
     "introspection", Introspection_test.suite;
+    "errors", Error_test.suite;
     "abstract", Abstract_test.suite;
   ]